+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/ChatInputHistory.kt
@@ -1,8 +1,0 @@
-@file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated
-
-data class ChatInputHistory(
-  val inputText: String? = null,
-  val inputContextFiles: List<ContextItem>? = null,
-)
-