+++ vscode/src/completions/detect-multiline.ts
@@ -16,7 +16,6 @@ import {
 interface DetectMultilineParams {
     docContext: LinesContext & DocumentDependentContext
     languageId: string
-    dynamicMultilineCompletions: boolean
     position: Position
 }
 
@@ -56,7 +55,7 @@ const LANGUAGES_WITH_MULTILINE_SUPPORT = [
 ]
 
 export function detectMultiline(params: DetectMultilineParams): DetectMultilineResult {
-    const { docContext, languageId, dynamicMultilineCompletions, position } = params
+    const { docContext, languageId, position } = params
     const { prefix, prevNonEmptyLine, nextNonEmptyLine, currentLinePrefix, currentLineSuffix } =
         docContext
     const isMultilineSupported = LANGUAGES_WITH_MULTILINE_SUPPORT.includes(languageId)
@@ -74,10 +73,9 @@ export function detectMultiline(params: DetectMultilineParams): DetectMultilineR
     // Don't fire multiline completion for method or function invocations
     // see https://github.com/sourcegraph/cody/discussions/358#discussioncomment-6519606
     // Don't fire multiline completion for unsupported languages.
-    if ((!dynamicMultilineCompletions && isMethodOrFunctionInvocation) || !isMultilineSupported) {
+    if (isMethodOrFunctionInvocation || !isMultilineSupported) {
         addAutocompleteDebugEvent('detectMultiline', {
             languageId,
-            dynamicMultilineCompletions,
             isMethodOrFunctionInvocation,
         })
 
@@ -106,9 +104,8 @@ export function detectMultiline(params: DetectMultilineParams): DetectMultilineR
         // than the block start line (the newly created block is empty).
         indentation(prevNonEmptyLine) >= indentation(nextNonEmptyLine)
 
-    if ((dynamicMultilineCompletions && isNewLineOpeningBracketMatch) || isSameLineOpeningBracketMatch) {
+    if (isNewLineOpeningBracketMatch || isSameLineOpeningBracketMatch) {
         addAutocompleteDebugEvent('detectMultiline', {
-            dynamicMultilineCompletions,
             isNewLineOpeningBracketMatch,
             isSameLineOpeningBracketMatch,
         })
@@ -120,7 +117,7 @@ export function detectMultiline(params: DetectMultilineParams): DetectMultilineR
     }
 
     const nonEmptyLineEndsWithBlockStart =
-        currentLinePrefix.length > 0 &&
+        currentLinePrefix.trim() !== '' &&
         isBlockStartActive &&
         indentation(currentLinePrefix) >= indentation(nextNonEmptyLine)
 
@@ -135,9 +132,8 @@ export function detectMultiline(params: DetectMultilineParams): DetectMultilineR
         // than the block start line (the newly created block is empty).
         indentation(prevNonEmptyLine) >= indentation(nextNonEmptyLine)
 
-    if ((dynamicMultilineCompletions && nonEmptyLineEndsWithBlockStart) || isEmptyLineAfterBlockStart) {
+    if (nonEmptyLineEndsWithBlockStart || isEmptyLineAfterBlockStart) {
         addAutocompleteDebugEvent('detectMultiline', {
-            dynamicMultilineCompletions,
             nonEmptyLineEndsWithBlockStart,
             isEmptyLineAfterBlockStart,
         })
@@ -149,7 +145,6 @@ export function detectMultiline(params: DetectMultilineParams): DetectMultilineR
     }
 
     addAutocompleteDebugEvent('detectMultiline', {
-        dynamicMultilineCompletions,
         nonEmptyLineEndsWithBlockStart,
         isEmptyLineAfterBlockStart,
         isNewLineOpeningBracketMatch,