+++ vscode/src/completions/get-inline-completions-tests/multiline-truncation.test.ts
@@ -756,7 +756,6 @@ for (const isTreeSitterEnabled of cases) {
                             }
                             console.log(5)┤`
                         },
-                        configuration: { autocompleteExperimentalDynamicMultilineCompletions: true },
                     })
 
                     const [insertText] = await getInlineCompletionsInsertText(requestParams)