+++ vscode/src/completions/providers/get-completion-params.ts
@@ -1,22 +1,16 @@
 import type { AutocompleteTimeouts, CodeCompletionsParams } from '@sourcegraph/cody-shared'
 
-import {
-    fetchAndProcessCompletions,
-    fetchAndProcessDynamicMultilineCompletions,
-} from './fetch-and-process-completions'
 import type { ProviderOptions } from './provider'
 
-const MAX_RESPONSE_TOKENS = 256
+export const MAX_RESPONSE_TOKENS = 256
 
-type LineNumberDependentCompletionParams = Pick<
-    CodeCompletionsParams,
-    'maxTokensToSample' | 'stopSequences' | 'timeoutMs'
+type LineNumberDependentCompletionParams = Required<
+    Pick<CodeCompletionsParams, 'maxTokensToSample' | 'stopSequences' | 'timeoutMs'>
 >
 
 interface LineNumberDependentCompletionParamsByType {
     singlelineParams: LineNumberDependentCompletionParams
     multilineParams: LineNumberDependentCompletionParams
-    dynamicMultilineParams: LineNumberDependentCompletionParams
 }
 
 interface Params {
@@ -31,30 +25,14 @@ export function getLineNumberDependentCompletionParams(
 
     return {
         singlelineParams: {
-            timeoutMs: 5_000,
-            // To speed up sample generation in single-line case, we request a lower token limit
-            // since we can't terminate on the first `\n`.
-            maxTokensToSample: 30,
+            timeoutMs: 15_000,
             stopSequences: singlelineStopSequences,
+            maxTokensToSample: MAX_RESPONSE_TOKENS,
         },
         multilineParams: {
             timeoutMs: 15_000,
-            maxTokensToSample: MAX_RESPONSE_TOKENS,
             stopSequences: multilineStopSequences,
-        },
-        dynamicMultilineParams: {
-            timeoutMs: 15_000,
             maxTokensToSample: MAX_RESPONSE_TOKENS,
-            // Do not stop after two consecutive new lines to get the full syntax node content. For example:
-            //
-            // function quickSort(array) {
-            //   if (array.length <= 1) {
-            //     return array
-            //   }
-            //
-            //   // the implementation continues here after two new lines.
-            // }
-            stopSequences: undefined,
         },
     }
 }
@@ -65,31 +43,22 @@ interface GetCompletionParamsAndFetchImplParams {
     timeouts?: AutocompleteTimeouts | undefined
 }
 
-interface GetRequestParamsAndFetchImplResult {
-    partialRequestParams: Omit<CodeCompletionsParams, 'messages'>
-    fetchAndProcessCompletionsImpl: typeof fetchAndProcessCompletions
-}
-
-export function getCompletionParamsAndFetchImpl(
+export function getCompletionParams(
     params: GetCompletionParamsAndFetchImplParams
-): GetRequestParamsAndFetchImplResult {
+): Omit<CodeCompletionsParams, 'messages'> {
     const {
         timeouts,
-        providerOptions: { multiline: isMutiline, dynamicMultilineCompletions, hotStreak },
-        lineNumberDependentCompletionParams: {
-            singlelineParams,
-            multilineParams,
-            dynamicMultilineParams,
-        },
+        providerOptions: { multiline: isMultiline, hotStreak },
+        lineNumberDependentCompletionParams: { singlelineParams, multilineParams },
     } = params
 
-    const useExtendedGeneration = isMutiline || dynamicMultilineCompletions || hotStreak
+    const useExtendedGeneration = isMultiline || hotStreak
 
-    const partialRequestParams: Omit<CodeCompletionsParams, 'messages'> = {
+    const partialRequestParams = {
         ...(useExtendedGeneration ? multilineParams : singlelineParams),
         temperature: 0.2,
         topK: 0,
-    }
+    } satisfies Omit<CodeCompletionsParams, 'messages'>
 
     // Apply custom multiline timeouts if they are defined.
     if (timeouts?.multiline && useExtendedGeneration) {
@@ -101,17 +70,5 @@ export function getCompletionParamsAndFetchImpl(
         partialRequestParams.timeoutMs = timeouts.singleline
     }
 
-    let fetchAndProcessCompletionsImpl = fetchAndProcessCompletions
-    if (dynamicMultilineCompletions) {
-        // If the feature flag is enabled use params adjusted for the experiment.
-        Object.assign(partialRequestParams, dynamicMultilineParams)
-
-        // Use an alternative fetch completions implementation.
-        fetchAndProcessCompletionsImpl = fetchAndProcessDynamicMultilineCompletions
-    }
-
-    return {
-        partialRequestParams,
-        fetchAndProcessCompletionsImpl,
-    }
+    return partialRequestParams
 }