+++ vscode/src/tree-sitter/parser.ts
@@ -3,7 +3,9 @@ import path from 'path'
 import * as vscode from 'vscode'
 import type Parser from 'web-tree-sitter'
 
-import { DOCUMENT_LANGUAGE_TO_GRAMMAR, type SupportedLanguage } from './grammars'
+import { wrapInActiveSpan } from '@sourcegraph/cody-shared'
+import type { Tree } from 'web-tree-sitter'
+import { DOCUMENT_LANGUAGE_TO_GRAMMAR, type SupportedLanguage, isSupportedLanguage } from './grammars'
 import { initQueries } from './query-sdk'
 const ParserImpl = require('web-tree-sitter') as typeof Parser
 
@@ -13,7 +15,7 @@ const ParserImpl = require('web-tree-sitter') as typeof Parser
  * and load language grammar only once, first time we need parser for a specific
  * language, next time we read it from this cache.
  */
-const PARSERS_LOCAL_CACHE: Partial<Record<SupportedLanguage, Parser>> = {}
+const PARSERS_LOCAL_CACHE: Partial<Record<SupportedLanguage, WrappedParser>> = {}
 
 interface ParserSettings {
     language: SupportedLanguage
@@ -25,7 +27,7 @@ interface ParserSettings {
     grammarDirectory?: string
 }
 
-export function getParser(language: SupportedLanguage): Parser | undefined {
+export function getParser(language: SupportedLanguage): WrappedParser | undefined {
     return PARSERS_LOCAL_CACHE[language]
 }
 
@@ -44,7 +46,14 @@ async function isRegularFile(uri: vscode.Uri): Promise<boolean> {
     }
 }
 
-export async function createParser(settings: ParserSettings): Promise<Parser | undefined> {
+export type WrappedParser = Pick<Parser, 'parse' | 'getLanguage'> & {
+    /**
+     * Wraps `parser.parse()` call into an OpenTelemetry span.
+     */
+    observableParse: Parser['parse']
+}
+
+export async function createParser(settings: ParserSettings): Promise<WrappedParser | undefined> {
     const { language, grammarDirectory = __dirname } = settings
 
     const cachedParser = PARSERS_LOCAL_CACHE[language]
@@ -54,6 +63,7 @@ export async function createParser(settings: ParserSettings): Promise<Parser | u
     }
 
     const wasmPath = path.resolve(grammarDirectory, DOCUMENT_LANGUAGE_TO_GRAMMAR[language])
+
     if (!(await isRegularFile(vscode.Uri.file(wasmPath)))) {
         return undefined
     }
@@ -64,12 +74,36 @@ export async function createParser(settings: ParserSettings): Promise<Parser | u
     const languageGrammar = await ParserImpl.Language.load(wasmPath)
 
     parser.setLanguage(languageGrammar)
-    // stop parsing after 50ms to avoid infinite loops
-    // if that happens, tree-sitter throws an error so we can catch and address it
-    parser.setTimeoutMicros(50_000)
-    PARSERS_LOCAL_CACHE[language] = parser
 
+    // Disable the timeout in unit tests to avoid timeout errors.
+    if (!process.env.VITEST) {
+        // Stop parsing after 70ms to avoid infinite loops.
+        // If that happens, tree-sitter throws an error so we can catch and address it.
+        parser.setTimeoutMicros(70_000)
+    }
+
+    const wrappedParser: WrappedParser = {
+        getLanguage: () => parser.getLanguage(),
+        parse: (...args) => parser.parse(...args),
+        observableParse: (...args) => wrapInActiveSpan('parser.parse', () => parser.parse(...args)),
+    }
+
+    PARSERS_LOCAL_CACHE[language] = wrappedParser
     initQueries(languageGrammar, language, parser)
 
-    return parser
+    return wrappedParser
+}
+
+export function parseString(languageId: string, source: string): Tree | null {
+    if (!isSupportedLanguage(languageId)) {
+        return null
+    }
+
+    const parser = getParser(languageId)
+
+    if (!parser) {
+        return null
+    }
+
+    return parser.parse(source)
 }